update:아카이브 리스트 갯수 표시 할때 무한호출 에러 수정 및 코드 재배치
